test: prune low-signal tests and consolidate test files
Remove 29 net-new tests plus delegation test file (~1100 lines) that
verified tautologies, library behaviour, compile-time invariants, or
one-line pass-throughs. Merge the 'more' and 'validation' split files
back into their single-file counterparts.

Deletions:
- limited_capabilities_delegation_test.go (whole file): verified
  stub.calls[method]==1 on pass-through methods. WriteReport gating
  logic (size+gas) remains covered by limited_capabilities_test.go.
- chaintype_test.go: literal-constant getters, crypto-library echo,
  duplicate ResolveKey cases already in the table, one-line
  delegation, compile-time interface check.
- health_more_test.go: nil/empty-map tautologies and exact duplicates
  of tests in health_test.go.
- supported_chains_test.go: pin tests (NotEmpty, EthereumMainnet,
  Sepolia with hardcoded forwarder), tautologies (Lowercased,
  Are20Bytes redundant with hex regex), compile-time struct check.
- trigger_validation_test.go: regex coverage redundant with parser
  table; misleading-name edge cases dropped; 5 genuinely novel cases
  (zero, max uint64, negative, unicode digits, tab) merged into the
  parser table.
- registry_test.go: map round-trip smokes and happy-path nil
  acceptance; all real branches covered by duplicate-panic, sort,
  defensive-copy, error-with-names, CLI-branching tests.

Consolidations:
- Merged health_more_test.go into health_test.go (11 tests total).
- Merged trigger_validation_test.go into trigger_test.go (8 tests
  plus the expanded parser table).

Kept tests exercise real branches and data invariants: ResolveKey
hex/sentinel logic, RegisterCapabilities type assertions, ExecuteTrigger
error paths, CollectCLIInputs filters, experimental/unknown health
labels, unique selectors, valid hex forwarders, chain-selector
resolution, httptest RPC round-trips in GetEVMTriggerLogFromValues,
and registry duplicate-panic / sort / defensive-copy / CLI branching.
